moved eff at cut code into working point script
