+++ ftag/tests/wps/test_working_points.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
+import tempfile
+from pathlib import Path
 import pytest
 
 from ftag.mock import get_mock_file
-from ftag.wps.working_points import get_working_points
+from ftag.wps.working_points import main
 
 
 @pytest.fixture
@@ -29,7 +31,7 @@ def test_get_working_points(ttbar_file, eff_val="60"):
         "-n",
         "10_000",
     ]
-    output = get_working_points(args)
+    output = main(args)
 
     assert "MockTagger" in output
     assert output["MockTagger"]["signal"] == "bjets"
@@ -59,7 +61,7 @@ def test_get_working_points_rejection(ttbar_file, rej_val="100"):
         "-r",
         "ujets",
     ]
-    output = get_working_points(args)
+    output =main(args)
 
     assert "MockTagger" in output
     assert output["MockTagger"]["signal"] == "bjets"
@@ -89,7 +91,7 @@ def test_get_working_points_cjets(ttbar_file, eff_val="60"):
         "-n",
         "10_000",
     ]
-    output = get_working_points(args)
+    output =main(args)
 
     assert "MockTagger" in output
     assert output["MockTagger"]["signal"] == "cjets"
@@ -119,7 +121,7 @@ def test_get_working_points_zprime(ttbar_file, zprime_file, eff_val="60"):
         "-n",
         "10_000",
     ]
-    output = get_working_points(args)
+    output =main(args)
 
     assert "MockTagger" in output
     assert output["MockTagger"]["signal"] == "bjets"
@@ -159,7 +161,7 @@ def test_get_working_points_xbb(ttbar_file, eff_val="60"):
         "hbb",  # Test for hbb signal
     ]
 
-    output = get_working_points(args)
+    output =main(args)
 
     assert "MockXbbTagger" in output
     assert output["MockXbbTagger"]["signal"] == "hbb"
@@ -177,8 +179,102 @@ def test_get_working_points_xbb(ttbar_file, eff_val="60"):
 def test_get_working_points_fx_length_check():
     # test with incorrect length of fx values for regular b-tagging
     with pytest.raises(ValueError):
-        get_working_points(["--ttbar", "path", "-t", "MockTagger", "-f", "0.1", "0.2"])
+       main(["--ttbar", "path", "-t", "MockTagger", "-f", "0.1", "0.2"])
 
     # test with incorrect length of fx values for Xbb tagging
     with pytest.raises(ValueError):
-        get_working_points(["--ttbar", "path", "--xbb", "-t", "MockXbbTagger", "-f", "0.25"])
+       main(["--ttbar", "path", "--xbb", "-t", "MockXbbTagger", "-f", "0.25"])
+
+
+def test_get_rej_eff_at_disc_ttbar(ttbar_file, disc_vals=None):
+    if disc_vals is None:
+        disc_vals = [1.0, 1.5]
+    args = [
+        "--ttbar",
+        str(ttbar_file),
+        "-t",
+        "MockTagger",
+        "-f",
+        "0.01",
+        "-n",
+        "10_000",
+    ]
+    args.extend(["-d"] + [str(x) for x in disc_vals])
+
+    output = main(args)
+
+    assert "MockTagger" in output
+    assert output["MockTagger"]["signal"] == "bjets"
+    assert output["MockTagger"]["fx"] == (0.01,)
+    assert "ttbar" in output["MockTagger"]
+    ttbar = output["MockTagger"]["ttbar"]
+    for dval in disc_vals:
+        assert str(dval) in ttbar
+        assert "eff" in ttbar[str(dval)]
+        assert "rej" in ttbar[str(dval)]
+
+    assert "zprime" not in output["MockTagger"]
+
+
+def test_get_rej_eff_at_disc_zprime(ttbar_file, zprime_file, disc_vals=None):
+    if disc_vals is None:
+        disc_vals = [1.0, 1.5]
+    args = [
+        "--ttbar",
+        str(ttbar_file),
+        "--zprime",
+        str(zprime_file),
+        "-t",
+        "MockTagger",
+        "-f",
+        "0.01",
+        "-n",
+        "10_000",
+    ]
+    args.extend(["-d"] + [str(x) for x in disc_vals])
+
+    output =main(args)
+
+    assert "MockTagger" in output
+    assert output["MockTagger"]["signal"] == "bjets"
+    assert output["MockTagger"]["fx"] == (0.01,)
+    assert "ttbar" in output["MockTagger"]
+    assert "zprime" in output["MockTagger"]
+
+    ttbar = output["MockTagger"]["ttbar"]
+    zprime = output["MockTagger"]["zprime"]
+
+    for dval in disc_vals:
+        for out in [ttbar, zprime]:
+            assert str(dval) in out
+            assert "eff" in out[str(dval)]
+            assert "rej" in out[str(dval)]
+
+
+def test_output_file(ttbar_file):
+    with tempfile.TemporaryDirectory() as tmpdir:
+        print(str(tmpdir))
+        output = str(tmpdir) + "/output.yaml"
+        args = [
+            "--ttbar",
+            str(ttbar_file),
+            "-t",
+            "MockTagger",
+            "-f",
+            "0.01",
+            "-n",
+            "10_000",
+            "-d",
+            "1.0",
+            "-o",
+            str(output),
+        ]
+
+        main(args)
+        assert Path(output).exists()
+
+
+def test_get_working_points_fx_length_check():
+    # test with incorrect length of fx values for regular b-tagging
+    with pytest.raises(ValueError):
+       main(["--ttbar", "path", "-t", "MockTagger", "-f", "0.1", "0.2", "-d", "1.0"])
